+++ tests/testthat/test-schema.R
@@ -1,7 +1,8 @@
 test_that("load_schema_table() works as expected", {
   # There are too many path dependencies to work on CI.
   skip_on_ci()
-  st <- load_schema_table(version = "v1.4")
+  skip_on_cran()
+  st <- load_schema_table(version = "v2.0")
   expect_true("Table" %in% colnames(st))
   expect_true("colname" %in% colnames(st))
   expect_true("coldatatype" %in% colnames(st))
@@ -13,7 +14,8 @@ test_that("load_schema_table() works as expected", {
 test_that("schema_dm() works as expected", {
   # There are too many path dependencies to work on CI.
   skip_on_ci()
-  clpfu_dm <- load_schema_table(version = "v1.4") |>
+  skip_on_cran()
+  clpfu_dm <- load_schema_table(version = "v2.0") |>
     schema_dm()
   clpfu_dm |>
     dm::dm_get_all_fks() |>
@@ -39,14 +41,16 @@ test_that("schema_dm() fails with unknown data type", {
 
 test_that("load_simple_tables() works as expected", {
   skip_on_ci()
-  simple_tables <- load_simple_tables(version = "v1.4")
+  skip_on_cran()
+  simple_tables <- load_simple_tables(version = "v2.0")
   expect_true("Year" %in% names(simple_tables))
   expect_true("Method" %in% names(simple_tables))
 })
 
 
-test_that("upload_schema_and_simple_tables() works as expected", {
+test_that("pl_upload_schema_and_simple_tables() works as expected", {
   skip_on_ci()
+  skip_on_cran()
   conn <- DBI::dbConnect(drv = RPostgres::Postgres(),
                          dbname = "unit_testing",
                          host = "eviz.cs.calvin.edu",
@@ -59,47 +63,15 @@ test_that("upload_schema_and_simple_tables() works as expected", {
 
   # Get the tables from the database
   tables <- DBI::dbListTables(conn)
-  for (table_name in c("Members", "Roles")) {
+  for (table_name in c("Member", "Role")) {
     expect_true(table_name %in% tables)
   }
-  members <- data.frame(Member_ID = as.integer(1:4),
+  members <- data.frame(MemberID = 1:4,
                         Member = c("John Lennon", "Paul McCartney", "George Harrison", "Ringo Starr"))
-  roles <- data.frame(Member_ID = as.integer(1:4),
+  roles <- data.frame(RoleID = 1:4,
                       Role = c("Lead singer", "Bassist", "Guitarist", "Drummer"))
-  expect_equal(DBI::dbReadTable(conn, "Members"), members)
-  expect_equal(DBI::dbReadTable(conn, "Roles"), roles)
+  expect_equal(DBI::dbReadTable(conn, "Member"), members)
+  expect_equal(DBI::dbReadTable(conn, "Role"), roles)
 })
 
 
-
-test_that("pl_upsert() works as expected", {
-  skip_on_ci()
-  conn <- DBI::dbConnect(drv = RPostgres::Postgres(),
-                         dbname = "unit_testing",
-                         host = "eviz.cs.calvin.edu",
-                         port = 5432,
-                         user = "postgres")
-  on.exit(DBI::dbDisconnect(conn))
-
-  # Build the data model remotely
-  PFUPipelineTools:::upload_beatles(conn)
-
-
-  # Try to upload more data for the fifth Beatle.
-  george_martin_member <- data.frame(Member_ID = as.integer(5),
-                                     Member = "George Martin")
-  george_martin_role <- data.frame(Member_ID = as.integer(5),
-                                   Role = "Producer")
-  # This should fail due to a bad primary key.
-  # There is no Name_ID = 5 in the Members table.
-  pl_upsert(george_martin_role, "Roles", conn) |>
-    expect_error('insert or update on table "Roles" violates foreign key constraint')
-  # Instead, add George Martin to the Members table so that his primary key will be available.
-  pl_upsert(george_martin_member, "Members", conn)
-  # Then Now add to the Roles table
-  pl_upsert(george_martin_role, "Roles", conn)
-  roles_tbl <- dplyr::tbl(conn, "Roles") |>
-    dplyr::collect()
-  expect_equal(nrow(roles_tbl), 5)
-  expect_equal(roles_tbl$Role, c("Lead singer", "Bassist", "Guitarist", "Drummer", "Producer"))
-})